Moved ArgValidation from ChartJs.Blazor.ChartJS.Common.Util to ChartJs.Blazor.Util.
